Module ipv4_heatmap::utils

Utility functions used by across the crate

Structs

BoundingBox

CIDRs in Hilbert space can represent a bounding box

Functions

bbox 🔒

Given the first (numeric) IP address in a CIDR block and the size of the CIDR block, return the bounding box. This handles the single point, square, and rectangle cases.

bbox_from_cidr

Given a CIDR in IP/## form, return the bounding box.

find_max

Find the maximum value in iterable

find_min

Find the minimum value in iterable

hil_xy_from_s

Convert an IPv4 address (in integer form) to a 12th order Hilbert x/y point

ip_to_numeric

Convert an characrter IPv4 address into an integer.

output_legend

Given a filename or path and palette name (+ whether the palette should be inverted) write an SVG legend out to the specified file.